Add comments to exercises/tests.sh
The final line in the shell script is exit 0. But why exit with exit code 0? The diff commands return a non-zero code for even the most trivial of diff's (such as a time stamp or a comment). We don't want the build system to fail for such trivial changes. So forcing an exit code of 0 allows the build to proceed.